% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\docType{package}
\name{ChrisFlow-package}
\alias{ChrisFlow-package}
\alias{ChrisFlow}
\title{TG Therapeutics Corporate Color Palettes}
\description{
Functions for working with TG Therapeutics corporate color palettes
in ggplot2 visualizations.
}
\author{
\strong{Maintainer}: Chris Rowland \email{crowland@tgtxinc.com} (Director, Clinical Data Science)

}
\keyword{internal}
