% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{export_to_excel}
\alias{export_to_excel}
\title{Export Data to Excel File}
\usage{
export_to_excel(
  data,
  file_name,
  output_folder = "output-files",
  col_names = TRUE,
  format_headers = TRUE
)
}
\arguments{
\item{data}{A data.table or data.frame to be exported.}

\item{file_name}{A string specifying the file name (without .xlsx extension).}

\item{output_folder}{A string specifying the output folder. Defaults to "output-files".}

\item{col_names}{Logical indicating whether to include column names. Defaults to TRUE.}

\item{format_headers}{Logical indicating whether to format headers. Defaults to TRUE.}
}
\value{
Invisibly returns the path to the saved file.
}
\description{
Exports a data.table or data.frame to an Excel file using writexl.
Automatically adds .xlsx extension and exports to "output-files" folder.
}
\examples{
\dontrun{
# Export a data.table to Excel
export_to_excel(mtcars, "car_data")
# Creates "output-files/car_data.xlsx"

# Export to a different folder
export_to_excel(mtcars, "car_data", output_folder = "results")
}

}
