% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_helpers.R
\name{hhmm_to_min}
\alias{hhmm_to_min}
\title{Convert HH:MM Duration to Minutes}
\usage{
hhmm_to_min(hhmm)
}
\arguments{
\item{hhmm}{An ITime object representing a duration in HH:MM format.}
}
\value{
Numeric value representing total minutes.
}
\description{
Converts ITime durations in HH:MM format to total minutes as a numeric value.
Useful for calculating duration statistics from time objects.
}
\examples{
\dontrun{
library(data.table)
hhmm_to_min(as.ITime("01:30"))
# Returns 90

hhmm_to_min(as.ITime("00:45"))
# Returns 45
}

}
