% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{import_access_tables_to_variables}
\alias{import_access_tables_to_variables}
\title{Import Microsoft Access Tables to Global Environment}
\usage{
import_access_tables_to_variables(connection, table_name)
}
\arguments{
\item{connection}{A connection object to the Microsoft Access Database
generated by \code{RODBC::odbcDriverConnect()}.}

\item{table_name}{A string specifying the name of the table to read, which will
also become the variable name (lowercased).}
}
\value{
This function is called for its side effect of assigning the data.table
to the global environment. Returns invisibly.
}
\description{
Reads tables from a Microsoft Access Database and assigns them to the global
environment with the name specified by the table_name argument. The data is
converted to a data.table and column names are cleaned using janitor.
}
\examples{
\dontrun{
# Create connection
conn <- RODBC::odbcDriverConnect("Driver={Microsoft Access Driver (*.mdb, *.accdb)};DBQ=path/to/db.accdb")

# Import table
import_access_tables_to_variables(connection = conn, table_name = "MyTable")
# The data is now available as `mytable` in the global environment
}

}
