% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{import_csv_files_to_variables}
\alias{import_csv_files_to_variables}
\title{Import CSV Files to Global Environment}
\usage{
import_csv_files_to_variables(folder, file_name)
}
\arguments{
\item{folder}{A string specifying the name of the folder containing the .csv file.}

\item{file_name}{A string specifying the name of the .csv file to read,
which will also become the variable name (lowercased, without .csv extension).}
}
\value{
This function is called for its side effect of assigning the data.table
to the global environment. Returns invisibly.
}
\description{
Reads a single .csv file in a specified folder and assigns it to the global
environment with the name derived from the file name. The data is read using
data.table::fread() and column names are cleaned using janitor.
}
\examples{
\dontrun{
import_csv_files_to_variables(folder = "data", file_name = "my_data.csv")
# The data is now available as `my_data` in the global environment
}

}
