% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{import_excel_files_to_variables}
\alias{import_excel_files_to_variables}
\title{Import Excel Files to Global Environment}
\usage{
import_excel_files_to_variables(folder, file_name, sheet)
}
\arguments{
\item{folder}{A string specifying the name of the folder containing the Excel file.}

\item{file_name}{A string specifying the name of the Excel file to read, which will
also become the variable name (lowercased, without .xlsx extension).}

\item{sheet}{A string specifying the name of the sheet to read.}
}
\value{
This function is called for its side effect of assigning the data.table
to the global environment. Returns invisibly.
}
\description{
Reads a single sheet from an Excel file in a specified folder and assigns it
to the global environment with the name derived from the file name. The data is
converted to a data.table and column names are cleaned using janitor.
}
\examples{
\dontrun{
import_excel_files_to_variables(
  folder = "data",
  file_name = "my_data.xlsx",
  sheet = "Sheet1"
)
# The data is now available as `my_data` in the global environment
}

}
