% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{import_sas_files_to_variables}
\alias{import_sas_files_to_variables}
\title{Import SAS Files to Global Environment}
\usage{
import_sas_files_to_variables(folder, file_name)
}
\arguments{
\item{folder}{A string specifying the name of the folder containing the SAS file.}

\item{file_name}{A string specifying the name of the .sas7bdat SAS file to read,
which will also become the variable name (lowercased, without .sas7bdat extension).}
}
\value{
This function is called for its side effect of assigning the data.table
to the global environment. Returns invisibly.
}
\description{
Reads a single .sas7bdat SAS file in a specified folder and assigns it to the
global environment with the name derived from the file name. The data is
converted to a data.table and column names are cleaned using janitor.
}
\examples{
\dontrun{
import_sas_files_to_variables(folder = "data", file_name = "my_data.sas7bdat")
# The data is now available as `my_data` in the global environment
}

}
