% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{import_xpt_files_to_variables}
\alias{import_xpt_files_to_variables}
\title{Import XPT Files to Global Environment}
\usage{
import_xpt_files_to_variables(folder, file_name)
}
\arguments{
\item{folder}{A string specifying the name of the folder containing the .xpt file.}

\item{file_name}{A string specifying the name of the .xpt file to read,
which will also become the variable name (lowercased, without .xpt extension).}
}
\value{
This function is called for its side effect of assigning the data.table
to the global environment. Returns invisibly.
}
\description{
Reads a single .xpt file in a specified folder and assigns it to the global
environment with the name derived from the file name. The data is converted
to a data.table and column names are cleaned using janitor.
}
\examples{
\dontrun{
import_xpt_files_to_variables(folder = "data", file_name = "my_data.xpt")
# The data is now available as `my_data` in the global environment
}

}
