% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_helpers.R
\name{last_day}
\alias{last_day}
\title{Get Last Day of Month}
\usage{
last_day(date)
}
\arguments{
\item{date}{A Date or IDate object.}
}
\value{
An IDate representing the last day of the month.
}
\description{
Returns the last day of the month for a given date.
}
\examples{
last_day(as.Date("2024-02-15"))
# Returns 2024-02-29 (leap year)

last_day(as.Date("2023-02-15"))
# Returns 2023-02-28

}
