% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{palette_gen}
\alias{palette_gen}
\title{Generate a Palette Function}
\usage{
palette_gen(palette = "main", direction = 1)
}
\arguments{
\item{palette}{Character name of desired palette, defaults to \code{"main"}}

\item{direction}{Numeric controlling the direction of the palette.
Use \verb{< 0} to reverse the order}
}
\value{
A function that takes \code{n} (number of colors) and returns a
character vector of hex color codes
}
\description{
This function outputs another function which accepts \code{n} as an argument.
The \code{n} refers to the number of colors needed for a particular plot.

Creates a discrete palette that will use the first \code{n} colors from
the supplied color values when the palette has enough colors.
Otherwise, uses an interpolated color palette via \code{colorRampPalette()}.

Can be used with \code{setNames()} to create named palettes so that specific
groups are always the same color.
}
\examples{
# Generate 3 colors from main palette
palette_gen()(3)

# Generate 5 colors from AE palette
palette_gen("ae")(5)

# Reverse palette direction
palette_gen("main", direction = -1)(3)

# Create named palette for treatment arms
arms <- c("ubli", "teri")
arms_named_palette <- setNames(
  object = palette_gen()(length(unique(arms))),
  nm = unique(arms)
)

}
