% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{ramp_hues}
\alias{ramp_hues}
\title{Generate Color Ramps from a Center Color}
\usage{
ramp_hues(center_color = NULL, ...)
}
\arguments{
\item{center_color}{A single color (hex code or color name) to act as
the center color being darkened and lightened}
}
\value{
A character vector of 7 hex color codes ranging from dark to light
}
\description{
Function that ramps from white to black through a center color,
outputting 7 colors. The output includes 50\% darker, 25\% darker,
center color, 25\% lighter, and 50\% lighter variations.
}
\examples{
ramp_hues("#99CC33")
ramp_hues("navy")

}
