% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{scale_fill_tg}
\alias{scale_fill_tg}
\title{TG Fill Scale for ggplot2}
\usage{
scale_fill_tg(palette = "main", discrete = TRUE, direction = 1, ...)
}
\arguments{
\item{palette}{Character name of palette in \code{tg_corp_palette()}.
Defaults to \code{"main"}}

\item{discrete}{Logical indicating whether fill aesthetic is discrete
or continuous. Defaults to \code{TRUE}}

\item{direction}{Numeric indicating whether the palette should be reversed.
Use \verb{< 0} to reverse. Defaults to \code{1}}

\item{...}{Additional arguments passed to \code{ggplot2::scale_fill_discrete()}
or \code{ggplot2::scale_fill_gradientn()}, used respectively when discrete
is TRUE or FALSE}
}
\value{
A ggplot2 scale object
}
\description{
Fill scale constructor for TG colors. Use this function to apply
TG corporate colors to the \code{fill} aesthetic in ggplot2.
}
\examples{
library(ggplot2)

# Discrete fill scale
ggplot(mtcars, aes(x = factor(cyl), fill = factor(cyl))) +
  geom_bar() +
  scale_fill_tg()

# Continuous fill scale
ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
  geom_tile() +
  scale_fill_tg(discrete = FALSE)

# Use highlight palette
ggplot(mtcars, aes(x = factor(am), fill = factor(am))) +
  geom_bar() +
  scale_fill_tg(palette = "highlight")

}
